Make sure device connection state is correct
This change is to overcome the scenario where another BLE2MQTT node has
connected to a peripheral before the current BLE2MQTT node has detected it was
disconnected or has not been able to publish an MQTT disconnect event yet.
The node which currently maintains the connection to the BLE peripheral will
make sure the connected state is correct.