Don't wait for ESP_GATTC_WRITE_DESCR_EVT
On some devices, when enabling notifications on the Services Changed
characteristics the ESP_GATTC_WRITE_DESCR_EVT event isn't received blocking the
BLE queue even though it was seen with a BLE sniffer that the peripheral did
respond to the write request.
As a workaround, we'll not wait for this event and simply dequeue the next
operation.