Try reconnecting to WiFi if MQTT fails 3 times
In some cases, WiFi connections seems to drop but the application doesn't get
notified about this for a while ~30 minutes. During this time it keeps trying to
connect to the MQTT broker but constantly times-out. When this happens, the
ESP32 will disconnect and reconnect to WiFi to overcome the issue.