Properly unsubscribe when devices disconnect
By the time we handle the BLE device disconnection event, it's already removed
from the devices list including all of the services and characteristic it
exposed. As such, we cannot traverse the characteristics anymore and
unsubscribe from each one specifically. In order to handle that, we now
unsubscribe to anything that started with the device's MAC address, i.e. any
MQTT subscription that related to the disconnected device.